+++ Mantids30/Net_Sockets/socket.h
@@ -231,7 +231,7 @@ public:
      * @param datalen data length in bytes
      * @return return the number of bytes read by the socket, zero for end of file and -1 for error.
      */
-    virtual ssize_t partialRead(void * data, const uint32_t & datalen);
+    virtual ssize_t partialRead(void * data, const size_t & datalen);
     /**
      * Write a data block to the socket
      * note that this haves some limitations. some systems can only send 4k at time.
@@ -240,7 +240,7 @@ public:
      * @param datalen data length in bytes
      * @return return the number of bytes read by the socket, zero for end of file and -1 for error.
      */
-    virtual ssize_t partialWrite(const void *data, const uint32_t & datalen);
+    virtual ssize_t partialWrite(const void *data, const size_t & datalen);
 
     /**
      * @brief iShutdown Internal protocol Shutdown
